pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whispercast"
version = "0.1.0" 
authors = [
    { name = "Mohamed Dhameem", email = "dhameem.study@example.com" },
]
description = "WhisperCast is a unified speech intelligence toolkit powered by OpenAI and Hugging Face Whisper models."
readme = "README.md"
license = "MIT" 
requires-python = ">=3.10"

classifiers = [
    "Development Status :: 3 - Alpha", # Or "4 - Beta", "5 - Production/Stable"
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "pandas",
    "python-dotenv",
    "openai",
    "transformers",
    "torch",
    "torchaudio",
    "librosa",
    "pyannote.audio>=3.1"
]

[project.urls]
Homepage = "https://github.com/mohameddhameem/whispercast" 
Repository = "https://github.com/mohameddhameem/whispercast"

# This makes your script runnable from the command line after installation
[project.scripts]
whispercast = "whispercast.cli:main"

# This tells setuptools where to find your package in the src layout
[tool.setuptools.packages.find]
where = ["src"]
include = ["whispercast*"] # Finds the whispercast package and any sub-packages
namespaces = false

# Example for optional dependencies (e.g., for development)
[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0"
]

# Optional: for including non-code files with your package
[tool.setuptools.package-data]
whispercast = ["sample_data/*.wav"]
